fix deploy contract
